requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0
captum>=0.6.0
transformers>=4.30.0

# Model interpretability
shap>=0.42.0
lime>=0.2.0.1
alibi>=0.7.0
alibi-detect>=0.11.0
imodels>=2.3.0
interpret>=0.4.0
pygam>=0.8.0

# Calibration and uncertainty
torchmetrics>=1.0.0
netcal>=1.3.0
properscoring>=0.1.0

# Fairness
fairlearn>=0.9.0
aequitas>=0.0.1

# Adversarial robustness
torchattacks>=3.4.0

# Tracking (optional)
wandb>=0.15.0
mlflow>=2.5.0

# Serving (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Development
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pytest-cov>=4.1.0

# Demo
streamlit>=1.25.0
gradio>=3.40.0
